requirements.txt
click
pyshp
shapely
xlrd
nltk
sklearn
gensim
pandas
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
geolocation-python
